The build works!
Objects wouldn't spawn because the indicator got the spawn object event before the ObjectSpawner got it. So InidicatorUI.IsVisible would be set to false and ObjectSpawner couldn't spawn objects when it is set to false. So I removed that unnecessary variable.

I also moved the indicator from the UI prefab to the Managers prefab under ObjectSpawner because ObjectSpawner needs to reference the indicator and it's better to store that in the prefab than manually setting the reference it in every level scene.